Implémentation réelle du GroundSceneNode pour qu'il affiche le monde effectivement généré

+++ modules/worldapi/world/IPointOfView.h
@@ -7,18 +7,18 @@ public:
 	virtual ~IPointOfView() = default;
 
 	virtual maths::vec3d getPosition() const = 0;
-	virtual double getHorizonDistance() const = 0;
+	virtual float getHorizonDistance() const = 0;
 };
 
 class PointOfView : public IPointOfView{
 public:
-	PointOfView(maths::vec3d position, double horizon)
+	PointOfView(maths::vec3d position, float horizon)
 		: _pos(position), _horizon(horizon) {}
 
 	maths::vec3d getPosition() const override { return _pos; }
-	double getHorizonDistance() const override { return _horizon; }
+	float getHorizonDistance() const override { return _horizon; }
 
 	maths::vec3d _pos;
-	double _horizon;
+	float _horizon;
 };
